Functional beta version
There are some bugs I haven't had time to fix yet, try if you can find them :)
Only the Dansk class is implemented, English class needs to be fixed for english to be an option. And I suggest we only write java doc for the LanguageHandler class, because all methods under that is all the same and called from there.